move core files into td.site, remove threat dragon package
