--- td.core/src/content/app.css
@@ -1,5 +1,0 @@
-@import "bootstrap";
-@import "toastr";
-@import "jointjs";
-@import "./joint-custom.css";
-@import "./spinner.css";
