+++ td.site/package.json
@@ -32,12 +32,17 @@
     "angular-route": "1.8.2",
     "angular-ui-bootstrap": "2.5.6",
     "angular-xeditable": "^0.10.0",
+    "bootstrap": "3.4.1",
     "font-awesome": "4.7.0",
+    "hotkeys-js": "^3.7.6",
+    "jointjs": "^2.2.1",
     "jquery": "^3.4.1",
+    "json-rules-engine": "3.1.0",
     "lodash": "^4.17.19",
-    "owasp-threat-dragon-core": "^1.4.0",
     "pug": "^3.0.1",
-    "pug-bootstrap": "0.0.16"
+    "pug-bootstrap": "0.0.16",
+    "rework-npm-cli": "^0.1.3",
+    "toastr": "2.1.4"
   },
   "devDependencies": {
     "angular-mocks": "1.8.0",
