+++ td.site/src/app/app.js
@@ -1,5 +1,4 @@
-﻿
-'use strict';
+﻿'use strict';
 
 var angular = require('angular');
 require('angular-ui-bootstrap');
@@ -7,7 +6,7 @@ require('angular-route');
 require('angular-xeditable');
 require('angular-animate');
 window.jQuery = require('jquery');
-require('owasp-threat-dragon-core');
+require('./core');
 
 //temporary fix for Chrome/Jointjs problem
 SVGElement.prototype.getTransformToElement = SVGElement.prototype.getTransformToElement || function (toElement) {
