+++ td.site/src/app/core/diagrams/StrideEditPane.html
