+++ td.site/src/app/core/diagrams/confirmReloadOnDirty.html
