+++ td.site/src/app/core/diagrams/diagramdirectives.js
