+++ td.site/src/app/core/diagrams/elementpropdirectives.js
