+++ td.site/src/content/app.css
@@ -1,6 +1,7 @@
-@import "owasp-threat-dragon-core";
+@import "bootstrap";
+@import "toastr";
+@import "jointjs";
+@import "./joint-custom.css";
+@import "./spinner.css";
 @import "angular-xeditable";
 @import "font-awesome";
-/*@import "./bootstrap-theme.css";*/
-/*@import "./bootstrap-custom.css";*/
-/*@import "./logo-nav.css";*/
